+++ app/src/main/java/com/barryburgle/gameapp/ui/input/card/EventCard.kt
@@ -24,15 +24,21 @@ import androidx.compose.material.icons.filled.PersonAddAlt1
 import androidx.compose.material.icons.filled.Share
 import androidx.compose.material.icons.filled.Stop
 import androidx.compose.material.icons.filled.Timer
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawWithContent
+import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.ClipboardManager
@@ -65,6 +71,8 @@ import com.barryburgle.gameapp.ui.input.card.body.SetBody
 import com.barryburgle.gameapp.ui.input.dialog.component.DialogTextComponent
 import com.barryburgle.gameapp.ui.input.dialog.leadName
 import com.barryburgle.gameapp.ui.input.liveSessionPulsingColor
+import com.barryburgle.gameapp.ui.tool.dialog.ConfirmButton
+import com.barryburgle.gameapp.ui.tool.dialog.DismissButton
 import com.barryburgle.gameapp.ui.utilities.button.IconShadowButton
 import com.barryburgle.gameapp.ui.utilities.text.body.LittleBodyText
 import com.barryburgle.gameapp.ui.utilities.text.body.MediumBodyText
@@ -91,6 +99,37 @@ fun EventCard(
     val uriHandler = LocalUriHandler.current
     var liveSessionTime: Long = 0
     var liveSessionLeads: Int = 0
+    var showDeleteConfirmDialog by remember { mutableStateOf(false) }
+
+    if (showDeleteConfirmDialog) {
+        AlertDialog(
+            containerColor = MaterialTheme.colorScheme.surfaceVariant,
+            modifier = Modifier.shadow(elevation = 10.dp),
+            onDismissRequest = {
+                showDeleteConfirmDialog = false
+            },
+            title = {
+                LargeTitleText(text = "Delete Live Session")
+            },
+            text = {
+                LittleBodyText("Stop Live Session without saving result?")
+            },
+            confirmButton = {
+                ConfirmButton {
+                    onEvent(GameEvent.DeleteSession(sortableGameEvent.event as AbstractSession))
+                    val intent = Intent(context, PersistentNotificationService::class.java)
+                    context.stopService(intent)
+                    showDeleteConfirmDialog = false
+                }
+            },
+            dismissButton = {
+                DismissButton {
+                    showDeleteConfirmDialog = false
+                }
+            }
+        )
+    }
+
     Card(
         modifier = modifier, colors = CardDefaults.cardColors(
             containerColor = MaterialTheme.colorScheme.surface
@@ -380,6 +419,9 @@ fun EventCard(
                                             PersistentNotificationService::class.java
                                         )
                                         context.stopService(intent)
+                                    },
+                                    onLongClick = {
+                                        showDeleteConfirmDialog = true
                                     },
                                     imageVector = Icons.Default.Stop,
                                     contentDescription = "Stop Live Session",